+++ Tests/AppKernel.php
@@ -29,6 +29,7 @@ class AppKernel extends Kernel {
 	public function registerBundles() {
 		return array(
 			new \Doctrine\Bundle\DoctrineBundle\DoctrineBundle(),
+			new \Doctrine\Bundle\DoctrineCacheBundle\DoctrineCacheBundle(),
 			new \Symfony\Bundle\AsseticBundle\AsseticBundle(),
 			new \Symfony\Bundle\FrameworkBundle\FrameworkBundle(),
 			new \Symfony\Bundle\TwigBundle\TwigBundle(),