+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 - BC breaks (follow `UPGRADE-2.0.md` to upgrade):
   - simplify the built-in form by replacing hidden fields with view variables
+- [#35]: added a cache for settings
+
+[#35]: https://github.com/craue/CraueConfigBundle/issues/35
 
 ## 1.4.2 (2017-03-27)